+++ public/r/8bit-resizable.json
@@ -1,0 +1,18 @@
+{
+  "$schema": "https://ui.shadcn.com/schema/registry-item.json",
+  "name": "8bit-resizable",
+  "type": "registry:component",
+  "title": "8-bit Resizable Panel",
+  "description": "Accessible resizable panel groups and layouts with keyboard support.",
+  "registryDependencies": [
+    "resizable"
+  ],
+  "files": [
+    {
+      "path": "components/ui/8bit/resizable.tsx",
+      "content": "\"use client\"\n\nimport * as ResizablePrimitive from \"react-resizable-panels\"\n\nimport { cn } from \"@/lib/utils\"\n\nconst ResizablePanelGroup = ({\n  className,\n  children,\n  ...props\n}: React.ComponentProps<typeof ResizablePrimitive.PanelGroup>) => (\n  <div className=\"relative\">\n    <ResizablePrimitive.PanelGroup\n      className={cn(\n        \"flex h-full w-full data-[panel-group-direction=vertical]:flex-col\",\n        className\n      )}\n      {...props}\n    >\n      {children}\n      <div className=\"absolute top-0 left-0 w-full h-1.5 bg-foreground dark:bg-ring pointer-events-none\" />\n      <div className=\"absolute bottom-0 w-full h-1.5 bg-foreground dark:bg-ring pointer-events-none\" />\n      <div className=\"absolute top-1 -left-1 w-1.5 h-1/2 bg-foreground dark:bg-ring pointer-events-none\" />\n      <div className=\"absolute bottom-1 -left-1 w-1.5 h-1/2 bg-foreground dark:bg-ring pointer-events-none\" />\n      <div className=\"absolute top-1 -right-1 w-1.5 h-1/2 bg-foreground dark:bg-ring pointer-events-none\" />\n      <div className=\"absolute bottom-1 -right-1 w-1.5 h-1/2 bg-foreground dark:bg-ring pointer-events-none\" />\n    </ResizablePrimitive.PanelGroup>\n  </div>\n)\n\nfunction ResizablePanel({\n  className,\n  children,\n  ...props\n}: React.ComponentProps<typeof ResizablePrimitive.Panel>) {\n  return (\n    <ResizablePrimitive.Panel className={cn(className)} {...props}>\n      {children}\n    </ResizablePrimitive.Panel>\n  )\n}\n\nconst ResizableHandle = ({\n  withHandle,\n  className,\n  ...props\n}: React.ComponentProps<typeof ResizablePrimitive.PanelResizeHandle> & {\n  withHandle?: boolean\n}) => (\n  <ResizablePrimitive.PanelResizeHandle\n    className={cn(\n      `relative flex w-[1px] items-center justify-center after:absolute after:inset-y-0 after:left-1/2 after:w-1 after:-translate-x-1/2 focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring focus-visible:ring-offset-1 data-[panel-group-direction=vertical]:h-px data-[panel-group-direction=vertical]:w-full data-[panel-group-direction=vertical]:after:left-0 data-[panel-group-direction=vertical]:after:h-1 data-[panel-group-direction=vertical]:after:w-full data-[panel-group-direction=vertical]:after:-translate-y-1/2 data-[panel-group-direction=vertical]:after:translate-x-0 [&[data-panel-group-direction=vertical]>div]:rotate-90`,\n      className\n    )}\n    {...props}\n  >\n    <div className=\"absolute w-1.5 h-full border border-dotted border-foreground dark:border-ring pointer-events-none\" />\n    {withHandle && (\n      <div className=\"relative bg-foreground dark:bg-ring w-full aspect-video h-[20px] z-10 border border-ring\"></div>\n    )}\n  </ResizablePrimitive.PanelResizeHandle>\n)\n\nexport { ResizablePanelGroup, ResizablePanel, ResizableHandle }\n",
+      "type": "registry:component",
+      "target": "components/ui/8bit/resizable.tsx"
+    }
+  ]
+}
